ENT-2923 - remove db access code from the verification thread pool
